+++ client/cache/bloom_filter_cache.go
@@ -1,3 +1,17 @@
+// Copyright 2014 beego Author. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package cache
 
 import (
@@ -12,43 +26,23 @@ import (
 type BloomFilterCache struct {
 	Cache
 	*bloom.BloomFilter
-	loadFunc       func(ctx context.Context, key string) (any, error)
-	expiration     time.Duration // set cache expiration, default never expire
-	notUpdateBloom bool          // update bloom key after put cache, default update
-}
-
-type BloomFilterCacheOption func(bfc *BloomFilterCache)
-
-func WithExpirationOpt(t time.Duration) BloomFilterCacheOption {
-	return func(bfc *BloomFilterCache) {
-		bfc.expiration = t
-	}
-}
-
-func WithUpdateBloomOpt(update bool) BloomFilterCacheOption {
-	return func(bfc *BloomFilterCache) {
-		bfc.notUpdateBloom = update
-	}
+	loadFunc   func(ctx context.Context, key string) (any, error)
+	expiration time.Duration // set cache expiration, default never expire
 }
 
 func NewBloomFilterCache(cache Cache, ln func(context.Context, string) (any, error), blm *bloom.BloomFilter,
-	opts ...BloomFilterCacheOption) (*BloomFilterCache, error) {
+	expiration time.Duration) (*BloomFilterCache, error) {
 
 	if cache == nil || ln == nil || blm == nil {
 		return nil, berror.Error(InvalidInitParameters, "missing required parameters")
 	}
 
-	bfc := &BloomFilterCache{
+	return &BloomFilterCache{
 		Cache:       cache,
 		BloomFilter: blm,
 		loadFunc:    ln,
-	}
-
-	for _, opt := range opts {
-		opt(bfc)
-	}
-
-	return bfc, nil
+		expiration:  expiration,
+	}, nil
 }
 
 func (bfc *BloomFilterCache) Get(ctx context.Context, key string) (any, error) {
@@ -70,15 +64,4 @@ func (bfc *BloomFilterCache) Get(ctx context.Context, key string) (any, error) {
 		}
 	}
 	return val, nil
-}
-
-func (bfc *BloomFilterCache) Put(ctx context.Context, key string, val any, expiration time.Duration) error {
-	err := bfc.Cache.Put(ctx, key, val, expiration)
-	if err != nil {
-		return err
-	}
-	if !bfc.notUpdateBloom {
-		bfc.AddString(key)
-	}
-	return nil
 }